+++ src/components/atoms/Input.jsx
@@ -1,12 +1,12 @@
 import React from "react";
 import { Input } from "@nextui-org/react";
 
-const InputAtomo = ({ type, placeholder, id, name, erros, register, icon, isReadOnly }) => {
+const InputAtomo = ({ type, placeholder, id, name, erros, register, icon, isReadOnly,defaultValue }) => {
   return (
     <div className="w-full flex flex-col gap-4">
 
       <div className="flex w-full h-full flex-wrap md:flex-nowrap mb-6 md:mb-0 gap-4 font-sans">
-        <Input isReadOnly={isReadOnly ? isReadOnly : false} size={"sm"} type={type} id={id} label={placeholder} name={name} endContent={icon}
+        <Input isReadOnly={isReadOnly ? isReadOnly : false} size={"sm"} type={type} id={id} label={placeholder} name={name} defaultValue={defaultValue ? defaultValue : ''} endContent={icon}
           {...register(name, {
             required: {
               value: true,
